requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.4
setuptools==74.0.0
six==1.16.0
sympy==1.13.2
timm==1.0.9
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
